Adding the code for proof of stake in the blockchain
The new blocks are decided by the amount of tokens that are being staked by different accounts.
The more tokens that are staked, the higher the likelihood of the next block being assigned to that node.

TODO: Make the blockchain clustered so that the clients can also mine their own blocks and appropriately assign them to nodes